requirements.txt
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==46.0.5
Django==6.0
django-crispy-forms==2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
fastapi-mail==1.6.2
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.32.1.post2
gpxpy==1.6.2
greenlet==3.3.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
passlib==1.7.4
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
regex==2025.11.3
reportlab==4.4.10
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
sqlparse==0.5.4
starlette==0.52.1
ttkbootstrap==1.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.8
